CofE: Fix defect. The speech name labels for Antipholus of Syracuse and Antipholus of Epheseus had all been broken into two parts. a) A defective extra  numbered line in the previous speech that held just 'ANTIPHOLUS' and b) a speech name line, unnumbered, that held just 'OF SYRACUSE' or 'OF EPHESEUS'.
Fixing the speech name lines was easy. The hard part was figuring out what to do with the extra numbered lines. I've chosen to leave them in the text, for now, but as comments. The line numbers don't actually appear in the MIT html as it is rendered, so the missing line numbers won't be obvious to anyone using the text as it currently stands. For those of us who want to make the line numbers show, though, the missing line numbers will at least be a source of questions. Some downstream consumers may want to renumber the text lines but I decided it was better for now to leave numbered lines the same as in the original file.